api/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
requests==2.31.0
python-dotenv==1.0.0
pyotp==2.9.0
flask-mail==0.9.1
flask-jwt-extended==4.5.0
flask-migrate==4.1.0